imdb/requirements.txt
asgiref==3.5.2
bcrypt==3.2.2
bootstrap4==0.1.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0
django-bootstrap3==21.2
django-braces==1.15.0
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-filter==22.1
django-heroku==0.3.1
django-shortuuidfield==0.1.3
django-toolbelt==0.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Faker==13.13.0
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.1
psycopg2==2.8.6
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
requests==2.28.1
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
stripe==3.5.0
tzdata==2022.1
urllib3==1.26.10
Werkzeug==2.1.2
whitenoise==6.2.0
wincertstore==0.2
zipp==3.8.0
